+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 - [#122] Dedupe items in the "extras" param if passed as a string, array, or set. ([@pdokas])
+- [#129] Clean `undefined` oauth params from the base string. ([@jeremyruppel])
 
 ## [v3.7.0] - 2018-02-14
 
@@ -187,6 +188,7 @@ This "release" marks the start of a complete rewrite of the Flickr SDK. Once the
 [#119]: https://github.com/flickr/flickr-sdk/pull/119
 [#121]: https://github.com/flickr/flickr-sdk/pull/121
 [#122]: https://github.com/flickr/flickr-sdk/pull/122
+[#129]: https://github.com/flickr/flickr-sdk/pull/129
 
 <!-- other links -->